Revert "Fixes for julia_version. Expand and consolidate julia_version tests. (#4151)"
This reverts commit fe3714e87161bf64d49b629af820a145e8a1df2a.
