+++ test/new.jl
@@ -7,6 +7,7 @@ using  Pkg.Resolve: ResolverError
 import Pkg.Artifacts: artifact_meta, artifact_path
 import Base.BinaryPlatforms: HostPlatform, Platform, platforms_match
 using  ..Utils
+import ..HistoricalStdlibVersions
 using Logging
 
 general_uuid = UUID("23338594-aafe-5451-b93e-139f81909106") # UUID for `General`
@@ -3074,6 +3075,143 @@ end
         end
     end
 end
+
+using Pkg.Types: is_stdlib
+@testset "is_stdlib() across versions" begin
+    HistoricalStdlibVersions.register!()
+
+    networkoptions_uuid = UUID("ca575930-c2e3-43a9-ace4-1e988b2c1908")
+    pkg_uuid = UUID("44cfe95a-1eb2-52ea-b672-e2afdf69b78f")
+
+    # Test NetworkOptions across multiple versions (It became an stdlib in v1.6+, and was registered)
+    @test is_stdlib(networkoptions_uuid)
+    @test is_stdlib(networkoptions_uuid, v"1.6")
+    @test !is_stdlib(networkoptions_uuid, v"1.5")
+    @test !is_stdlib(networkoptions_uuid, v"1.0.0")
+    @test !is_stdlib(networkoptions_uuid, v"0.7")
+    @test !is_stdlib(networkoptions_uuid, nothing)
+
+    # Pkg is an unregistered stdlib and has always been an stdlib
+    @test is_stdlib(pkg_uuid)
+    @test is_stdlib(pkg_uuid, v"1.0")
+    @test is_stdlib(pkg_uuid, v"1.6")
+    @test is_stdlib(pkg_uuid, v"999.999.999")
+    @test is_stdlib(pkg_uuid, v"0.7")
+    @test is_stdlib(pkg_uuid, nothing)
+
+    HistoricalStdlibVersions.unregister!()
+    # Test that we can probe for stdlibs for the current version with no STDLIBS_BY_VERSION,
+    # but that we throw a PkgError if we ask for a particular julia version.
+    @test is_stdlib(networkoptions_uuid)
+    @test_throws Pkg.Types.PkgError is_stdlib(networkoptions_uuid, v"1.6")
+end
+
+
+@testset "Pkg.add() with julia_version" begin
+    HistoricalStdlibVersions.register!()
+
+    # A package with artifacts that went from normal package -> stdlib
+    gmp_jll_uuid = "781609d7-10c4-51f6-84f2-b8444358ff6d"
+    # A package that has always only ever been an stdlib
+    linalg_uuid = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+    # A package that went from normal package - >stdlib
+    networkoptions_uuid = "ca575930-c2e3-43a9-ace4-1e988b2c1908"
+
+    function get_manifest_block(name)
+        manifest_path = joinpath(dirname(Base.active_project()), "Manifest.toml")
+        @test isfile(manifest_path)
+        deps = Base.get_deps(TOML.parsefile(manifest_path))
+        @test haskey(deps, name)
+        return only(deps[name])
+    end
+
+    isolate(loaded_depot=true) do
+        # Next, test that if we ask for `v1.5` it DOES have a version, and that GMP_jll installs v6.1.X
+        Pkg.add(["NetworkOptions", "GMP_jll"]; julia_version=v"1.5")
+        no_block = get_manifest_block("NetworkOptions")
+        @test haskey(no_block, "uuid")
+        @test no_block["uuid"] == networkoptions_uuid
+        @test haskey(no_block, "version")
+
+        gmp_block = get_manifest_block("GMP_jll")
+        @test haskey(gmp_block, "uuid")
+        @test gmp_block["uuid"] == gmp_jll_uuid
+        @test haskey(gmp_block, "version")
+        @test startswith(gmp_block["version"], "6.1.2")
+
+        # Test that the artifact of GMP_jll contains the right library
+        @test haskey(gmp_block, "git-tree-sha1")
+        gmp_jll_dir = Pkg.Operations.find_installed("GMP_jll", Base.UUID(gmp_jll_uuid), Base.SHA1(gmp_block["git-tree-sha1"]))
+        @test isdir(gmp_jll_dir)
+        artifacts_toml = joinpath(gmp_jll_dir, "Artifacts.toml")
+        @test isfile(artifacts_toml)
+        meta = artifact_meta("GMP", artifacts_toml)
+
+        # `meta` can be `nothing` on some of our newer platforms; we _know_ this should
+        # not be the case on the following platforms, so we check these explicitly to
+        # ensure that we haven't accidentally broken something, and then we gate some
+        # following tests on whether or not `meta` is `nothing`:
+        for arch in ("x86_64", "i686"), os in ("linux", "mac", "windows")
+            if platforms_match(HostPlatform(), Platform(arch, os))
+                @test meta !== nothing
+            end
+        end
+
+        # These tests require a matching platform artifact for this old version of GMP_jll,
+        # which is not the case on some of our newer platforms.
+        if meta !== nothing
+            gmp_artifact_path = artifact_path(Base.SHA1(meta["git-tree-sha1"]))
+            @test isdir(gmp_artifact_path)
+
+            # On linux, we can check the filename to ensure it's grabbing the correct library
+            if Sys.islinux()
+                libgmp_filename = joinpath(gmp_artifact_path, "lib", "libgmp.so.10.3.2")
+                @test isfile(libgmp_filename)
+            end
+        end
+    end
+
+    # Next, test that if we ask for `v1.6`, GMP_jll gets `v6.2.0`, and for `v1.7`, it gets `v6.2.1`
+    function do_gmp_test(julia_version, gmp_version)
+        isolate(loaded_depot=true) do
+            Pkg.add("GMP_jll"; julia_version)
+            gmp_block = get_manifest_block("GMP_jll")
+            @test haskey(gmp_block, "uuid")
+            @test gmp_block["uuid"] == gmp_jll_uuid
+            @test haskey(gmp_block, "version")
+            @test startswith(gmp_block["version"], string(gmp_version))
+        end
+    end
+    do_gmp_test(v"1.6", v"6.2.0")
+    do_gmp_test(v"1.7", v"6.2.1")
+
+    isolate(loaded_depot=true) do
+        # Next, test that if we ask for `nothing`, NetworkOptions has a `version` but `LinearAlgebra` does not.
+        Pkg.add(["LinearAlgebra", "NetworkOptions"]; julia_version=nothing)
+        no_block = get_manifest_block("NetworkOptions")
+        @test haskey(no_block, "uuid")
+        @test no_block["uuid"] == networkoptions_uuid
+        @test haskey(no_block, "version")
+        linalg_block = get_manifest_block("LinearAlgebra")
+        @test haskey(linalg_block, "uuid")
+        @test linalg_block["uuid"] == linalg_uuid
+        @test !haskey(linalg_block, "version")
+    end
+
+    isolate(loaded_depot=true) do
+        # Next, test that stdlibs do not get dependencies from the registry
+        # NOTE: this test depends on the fact that in Julia v1.6+ we added
+        # "fake" JLLs that do not depend on Pkg while the "normal" p7zip_jll does.
+        # A future p7zip_jll in the registry may not depend on Pkg, so be sure
+        # to verify your assumptions when updating this test.
+        Pkg.add("p7zip_jll")
+        p7zip_jll_uuid = UUID("3f19e933-33d8-53b3-aaab-bd5110c3b7a0")
+        @test !("Pkg" in keys(Pkg.dependencies()[p7zip_jll_uuid].dependencies))
+    end
+
+    HistoricalStdlibVersions.unregister!()
+end
+
 
 @testset "Issue #2931" begin
     isolate(loaded_depot=false) do
