+++ test/runtests.jl
@@ -20,6 +20,7 @@ if realpath(dirname(dirname(Base.pathof(Pkg)))) != realpath(dirname(@__DIR__))
 end
 
 ENV["JULIA_PKG_PRECOMPILE_AUTO"]=0
+ENV["HISTORICAL_STDLIB_VERSIONS_AUTO_REGISTER"]="false"
 
 logdir = get(ENV, "JULIA_TEST_VERBOSE_LOGS_DIR", nothing)
 ### Send all Pkg output to a file called Pkg.log
@@ -35,6 +36,29 @@ end
 
 include("utils.jl")
 Logging.with_logger((islogging || Pkg.DEFAULT_IO[] == devnull) ? Logging.ConsoleLogger(Pkg.DEFAULT_IO[]) : Logging.current_logger()) do
+
+    # Because julia CI doesn't run stdlib tests via `Pkg.test` test deps must be manually installed if missing
+    if Base.find_package("HistoricalStdlibVersions") === nothing
+        @debug "Installing HistoricalStdlibVersions for Pkg tests"
+        iob = IOBuffer()
+        Pkg.activate(; temp = true)
+        try
+            # Needed for custom julia version resolve tests
+            # Don't use the toplevel PKg.add() command to avoid accidentally installing another copy of the registry
+            spec = Pkg.PackageSpec(
+                name="HistoricalStdlibVersions", 
+                url="https://github.com/JuliaPackaging/HistoricalStdlibVersions.jl",
+                rev="5879c5f690795208481c60b904f4af4e8c1eeef8", #= version="2.0.0", =#
+                uuid="6df8b67a-e8a0-4029-b4b7-ac196fe72102")
+            Pkg.API.handle_package_input!(spec)
+            Pkg.add(Pkg.API.Context(), [spec], io=iob)
+        catch
+            println(String(take!(iob)))
+            rethrow()
+        end
+    end
+
+    @eval import HistoricalStdlibVersions
 
     if (server = Pkg.pkg_server()) !== nothing && Sys.which("curl") !== nothing
         s = read(`curl -sLI $(server)`, String);
@@ -66,10 +90,6 @@ Logging.with_logger((islogging || Pkg.DEFAULT_IO[] == devnull) ? Logging.Console
         ]
 
     # Only test these if the test deps are available (they aren't typically via `Base.runtests`)
-    HSV_pkgid = Base.PkgId(Base.UUID("6df8b67a-e8a0-4029-b4b7-ac196fe72102"), "HistoricalStdlibVersions")
-    if Base.locate_package(HSV_pkgid) !== nothing
-        push!(test_files, "historical_stdlib_version.jl")
-    end
     Aqua_pkgid = Base.PkgId(Base.UUID("4c88cf16-eb10-579e-8560-4a9242c79595"), "Aqua")
     if Base.locate_package(Aqua_pkgid) !== nothing
         push!(test_files, "aqua.jl")
